сдача 16.05

+++ output.cpp
@@ -4,7 +4,7 @@ Output::Output()
 {
 
 }
-void Output::size_changed_print(QString name, qint64 size)
+void Output::size_changed_print(QString name, int size)
 {
      QTextStream out(stdout);
      out<<"File "<<name<<" changed. Size = "<<size<<endl;
@@ -16,7 +16,7 @@ void Output::file_deleted_print(QString name)
      out<<"File "<<name<<" deleted"<<endl;
 
 }
-void Output::file_created_print(QString name, qint64 size)
+void Output::file_created_print(QString name, int size)
 {
      QTextStream out(stdout);
      out<<"File "<<name<<" created. Size = "<<size<<endl;
