+++ output.h
@@ -10,9 +10,9 @@ class Output : public QObject
 public:
     Output();
 public slots:
-    void size_changed_print(QString name, qint64 size);
+    void size_changed_print(QString name, int size);
     void file_deleted_print(QString name);
-    void file_created_print(QString name, qint64 size);
+    void file_created_print(QString name, int size);
 };
 
 #endif // OUTPUT_H
